Fix the cast of alpha

+++ dsparse/lib/core_zsytrfsp.c
@@ -84,11 +84,11 @@ static void core_zsytf2sp(dague_int_t  n,
         cblas_zscal(n-k-1, CBLAS_SADDR( alpha ), tmp1, 1 );
 
         alpha = -(*tmp);
+
         /* TODO: replace by SYR but [cz]syr exists only in LAPACK */
-        LAPACKE_zsyr_work(LAPACK_COL_MAJOR, 'l',
-                          n-k-1, (double)(alpha),
-                          tmp1,        1,
-                          tmp1+stride, stride);
+        LAPACKE_zsyr_work(LAPACK_COL_MAJOR, 'l', n-k-1, 
+                          alpha, tmp1,        1,
+                                 tmp1+stride, stride);
     }
 }
